+++ src/services/event.js
@@ -1,4 +1,0 @@
-function submitForm() {
-    const searchEngine = document.getElementById("search-engine");
-    setValue("Engine", searchEngine.value)
-}