+++ Changelog.md
@@ -1,9 +1,15 @@
 # Change Log
 
-## Version 1.0.0
-* Added Search Page
+## Version 1.0.2 | 6/8/2022
+* Change Famework to **React**
+* Fixed Search not working on tab
+* Change some UI
+* Removed Settings Because of Error
 
-## Version 1.0.1
+## Version 1.0.1 | 5/8/2022
 * Added Builded Zip file
 * Added Settings
 * Added Custom Google, Bing, DuckDuckGo Search Engine Support
+
+## Version 1.0.0 | 4/8/2022
+* The Beginning?